refactor(interview): consume stage images directly; drop Architect nav toggle
