+++ apps/architect-web/src/components/PreviewHost/launchPreview.ts
@@ -12,7 +12,6 @@ type LaunchOptions = {
   startStage: number;
   useSyntheticData: boolean;
   skipLogicBypassed: boolean;
-  allowStageNavigation: boolean;
 };
 
 type LaunchPreviewResult = { kind: 'delivered' } | { kind: 'popup-blocked' };
@@ -22,7 +21,6 @@ export function launchPreview({
   startStage,
   useSyntheticData,
   skipLogicBypassed,
-  allowStageNavigation,
 }: LaunchOptions): Promise<LaunchPreviewResult> {
   const protocolId = getActiveProtocolScope();
   if (!protocolId) {
@@ -56,7 +54,6 @@ export function launchPreview({
     startStage,
     useSyntheticData,
     skipLogicBypassed,
-    allowStageNavigation,
   };
 
   return new Promise<LaunchPreviewResult>((resolve, reject) => {
