+++ apps/architect-web/src/ducks/modules/__tests__/app.test.ts
@@ -2,9 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import { describe, expect, it } from 'vitest';
 
 import appReducer, {
-  getPreviewAllowStageNavigation,
   getPreviewUseSyntheticData,
-  setPreviewAllowStageNavigation,
   setPreviewUseSyntheticData,
 } from '../app';
 
@@ -29,16 +27,5 @@ describe('app slice — preview preferences', () => {
     store.dispatch(setPreviewUseSyntheticData(false));
     store.dispatch(setPreviewUseSyntheticData(true));
     expect(getPreviewUseSyntheticData(store.getState())).toBe(true);
-  });
-
-  it('getPreviewAllowStageNavigation defaults to false when unset', () => {
-    const store = createStore();
-    expect(getPreviewAllowStageNavigation(store.getState())).toBe(false);
-  });
-
-  it('setPreviewAllowStageNavigation(true) flips the preference', () => {
-    const store = createStore();
-    store.dispatch(setPreviewAllowStageNavigation(true));
-    expect(getPreviewAllowStageNavigation(store.getState())).toBe(true);
   });
 });
