+++ packages/interview/package.json
@@ -48,6 +48,7 @@
   },
   "dependencies": {
     "@base-ui/react": "catalog:",
+    "@codaco/interface-images": "workspace:*",
     "@codaco/network-query": "workspace:^",
     "@faker-js/faker": "catalog:",
     "@mapbox/search-js-react": "^1.5.1",
@@ -69,7 +70,6 @@
   },
   "devDependencies": {
     "@chromatic-com/storybook": "catalog:",
-    "@codaco/interface-images": "workspace:*",
     "@codaco/protocol-utilities": "workspace:^",
     "@codaco/tsconfig": "workspace:*",
     "@playwright/test": "catalog:",
